Python/hw4/src/requirements.txt
autopep8==1.5.2
backcall==0.1.0
decorator==4.4.2
Django==2.2.12
django-debug-toolbar==2.2
django-extensions==2.2.9
entrypoints==0.3
Faker==4.0.3
flake8==3.7.9
flake8-builtins==1.5.2
flake8-import-order==0.18.1
flake8-print==3.1.4
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
wcwidth==0.1.9
